Setup automatic build and deploy via travis
This is a forked repo from github.com/kubernetes-incubator/external-storage
openebs-build.sh will copy this repository contents to the correct
  gopath and build the binaries.
openebs-deploy.sh will push the image to the dockerhub.
.travis.yml has been modified to run only on the release branch
  and build and deploy openebs-provisioner.

The builds are done on release branch instead of master, to help
keep the master in sync with the upstream/orig repository. The changes
from forked/master and sync from orig/master, that impact openebs-provisioner
will have to be merged to 'release' branch.